aduc_sdr/requirements.txt
torch
torchao
torchvision
torchaudio
transformers
accelerate
safetensors
einops
sentencepiece
git+https://github.com/hkchengrex/MMAudio.git@main
git+https://github.com/huggingface/diffusers.git@main
gradio>=5.23.1
gradio[oauth]
fastapi
uvicorn[standard]
pydantic
soundfile
tiktoken
transformers_stream_generator
rotary-embedding-torch
Pillow
numpy
PyYAML
opencv-python
imageio
imageio-ffmpeg
av
huggingface_hub
google-generativeai
torchmetrics            
pycocotools           
torch-fidelity         
moviepy              
imageio                
tabulate               
deepdiff             
parameterized         
mediapy             
black             
flake8             
isort                
pre-commit               
expecttest           
hypothesis           
